* undid change to migration * minor tidying of order spec * removes product routes spec * refactors methods in product controllers * completes testing of product routes/controllers

+++ migrations/sqls/20220309043741-products-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products (
     id SERIAL PRIMARY KEY,
     name varchar(100) NOT NULL,
-    price float(2) DEFAULT 0.00,
+    price money DEFAULT 0.00,
     category varchar(50)
 )
